Move protocol schemas to root Protocols namespace
