+++ src/AvroSourceGenerator.Core/Protocols/ProtocolRequestParameter.cs
@@ -1,6 +1,7 @@
-﻿using System.Text.Json;
+using System.Text.Json;
+using AvroSourceGenerator.Schemas;
 
-namespace AvroSourceGenerator.Schemas;
+namespace AvroSourceGenerator.Protocols;
 
 public sealed record class ProtocolRequestParameter(
     string Name,
@@ -31,3 +32,4 @@ public sealed record class ProtocolRequestParameter(
         writer.WriteEndObject();
     }
 }
+
