+++ src/AvroSourceGenerator.Core/Protocols/ProtocolResponse.cs
@@ -1,6 +1,7 @@
-﻿using System.Text.Json;
+using System.Text.Json;
+using AvroSourceGenerator.Schemas;
 
-namespace AvroSourceGenerator.Schemas;
+namespace AvroSourceGenerator.Protocols;
 
 public sealed record class ProtocolResponse(
     AvroSchema Type,
@@ -10,3 +11,4 @@ public sealed record class ProtocolResponse(
     public void WriteTo(Utf8JsonWriter writer, IReadOnlyDictionary<SchemaName, TopLevelSchema> registeredSchemas, HashSet<SchemaName> writtenSchemas, string? containingNamespace) =>
         Type.WriteTo(writer, registeredSchemas, writtenSchemas, containingNamespace);
 }
+
