+++ src/AvroSourceGenerator.Core/Protocols/ProtocolSchema.cs
@@ -1,7 +1,8 @@
-﻿using System.Collections.Immutable;
+using System.Collections.Immutable;
 using System.Text.Json;
+using AvroSourceGenerator.Schemas;
 
-namespace AvroSourceGenerator.Schemas;
+namespace AvroSourceGenerator.Protocols;
 
 public sealed record class ProtocolSchema(
     JsonElement Json,
@@ -58,3 +59,4 @@ public sealed record class ProtocolSchema(
         writer.WriteEndObject();
     }
 }
+
