+++ src/AvroSourceGenerator.Core/Registry/SchemaRegistry.Protocol.Messages.cs
@@ -1,6 +1,7 @@
-﻿using System.Collections.Immutable;
+using System.Collections.Immutable;
 using System.Text.Json;
 using AvroSourceGenerator.Extensions;
+using AvroSourceGenerator.Protocols;
 using AvroSourceGenerator.Schemas;
 
 namespace AvroSourceGenerator.Registry;
@@ -25,3 +26,4 @@ public readonly partial struct SchemaRegistry
         return new ProtocolMessage(methodName, documentation, requestParameters, response, errors);
     }
 }
+
