+++ src/AvroSourceGenerator.Core/Registry/SchemaRegistry.Protocol.RequestParameters.cs
@@ -1,6 +1,7 @@
-﻿using System.Collections.Immutable;
+using System.Collections.Immutable;
 using System.Text.Json;
 using AvroSourceGenerator.Extensions;
+using AvroSourceGenerator.Protocols;
 using AvroSourceGenerator.Schemas;
 
 namespace AvroSourceGenerator.Registry;
@@ -34,3 +35,4 @@ public readonly partial struct SchemaRegistry
         return new ProtocolRequestParameter(name, type, underlyingType, isNullable, documentation, defaultJson, @default);
     }
 }
+
