+++ src/AvroSourceGenerator.Core/Registry/SchemaRegistry.Protocol.cs
@@ -1,6 +1,7 @@
-﻿using System.Collections.Immutable;
+using System.Collections.Immutable;
 using System.Text.Json;
 using AvroSourceGenerator.Extensions;
+using AvroSourceGenerator.Protocols;
 using AvroSourceGenerator.Schemas;
 
 namespace AvroSourceGenerator.Registry;
@@ -27,3 +28,4 @@ public readonly partial struct SchemaRegistry
         }
     }
 }
+
